requirements.txt
flask
geopy
gpxpy
jinja2
numpy
pytz
pyuwsgi
pyyaml
requests
schema
scipy
timezonefinder
uwsgidecorators
